Add method to play five-round game in the console
Add a method to play a five-round game in the browser console. At the
end of each round, the winner of the round gets a point. If there's a
tie, noone gets a point and the game continues to the next round. At the
end of five rounds, the winner is declared based on the final score (if
the scores are tied, a tie is declared). Users input their answers into
a prompt. Users can't go to the next round if an invalid answer is given
(i.e. no variation of "rock", "paper", or "scissors").