Add UI and restructure file system
Add UI to the game so that the player can play with buttons instead of
with the browser console. Change the endgame rules so that the first player
to get five points wins the game. Restructure the file system by moving
the javascript code into a "scripts" folder and putting images into an
"images" folder.